+++ example/blog_plugins.py
@@ -1,14 +1,12 @@
 """
-This example demonstrates the use of caches
+This example demonstrates the use of plugins
 """
-from flask import Flask, render_template_string, redirect, current_app
-from sqlalchemy import create_engine, MetaData
-from flask_login import UserMixin, LoginManager, login_user, logout_user, current_user
-from flask_blogging import SQLAStorage, BloggingEngine
-from flask_principal import identity_changed, Identity, AnonymousIdentity, identity_loaded, \
-    UserNeed, RoleNeed
-from flask.ext.cache import Cache
 
+
+from flask import Flask, render_template_string, redirect
+from sqlalchemy import create_engine, MetaData
+from flask_login import UserMixin, LoginManager, login_user, logout_user
+from flask_blogging import SQLAStorage, BloggingEngine
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "secret"  # for WTF-forms and login
@@ -16,19 +14,17 @@ app.config["BLOGGING_URL_PREFIX"] = "/blog"
 app.config["BLOGGING_DISQUS_SITENAME"] = "test"
 app.config["BLOGGING_SITEURL"] = "http://localhost:8000"
 app.config["BLOGGING_SITENAME"] = "My Site"
-app.config["BLOGGING_PERMISSIONS"] = False  # Enable blogger permissions'
-app.config["CACHE_TYPE"] = "simple"
-
-# create cache
-cache = Cache(app)
+app.config["BLOGGING_PLUGINS"] = ["example.plugins.add_view",
+                                  "example.plugins.tag_cloud"]
 
 # extensions
 engine = create_engine('sqlite:////tmp/blog.db')
 meta = MetaData()
 sql_storage = SQLAStorage(engine, metadata=meta)
-blog_engine = BloggingEngine(app, sql_storage, cache=cache)
+blog_engine = BloggingEngine(app, sql_storage)
 login_manager = LoginManager(app)
 meta.create_all(bind=engine)
+
 
 class User(UserMixin):
     def __init__(self, user_id):
@@ -36,15 +32,6 @@ class User(UserMixin):
 
     def get_name(self):
         return "Paul Dirac"  # typically the user's name
-
-
-@identity_loaded.connect_via(app)
-def on_identity_loaded(sender, identity):
-    identity.user = current_user
-    if hasattr(current_user, "id"):
-        identity.provides.add(UserNeed(current_user.id))
-    identity.provides.add(RoleNeed("blogger"))
-
 
 @login_manager.user_loader
 @blog_engine.user_loader
@@ -76,21 +63,13 @@ def index():
 def login():
     user = User("testuser")
     login_user(user)
-
-    # notify the change of role
-    identity_changed.send(current_app._get_current_object(),
-                          identity=Identity("testuser"))
     return redirect("/blog")
 
 @app.route("/logout/")
 def logout():
     logout_user()
-
-    # notify the change of role
-    identity_changed.send(current_app._get_current_object(),
-                          identity=AnonymousIdentity())
     return redirect("/")
 
 
 if __name__ == "__main__":
-    app.run(debug=True, port=8000, use_reloader=True)
+    app.run(debug=True, port=8000, use_reloader=False)