+++ example/blog_roles.py
@@ -1,8 +1,8 @@
 from flask import Flask, render_template_string, redirect, current_app
 from sqlalchemy import create_engine, MetaData
-from flask.ext.login import UserMixin, LoginManager, login_user, logout_user, current_user
-from flask.ext.blogging import SQLAStorage, BloggingEngine
-from flask.ext.principal import identity_changed, Identity, AnonymousIdentity, identity_loaded, \
+from flask_login import UserMixin, LoginManager, login_user, logout_user, current_user
+from flask_blogging import SQLAStorage, BloggingEngine
+from flask_principal import identity_changed, Identity, AnonymousIdentity, identity_loaded, \
     UserNeed, RoleNeed
 
 
@@ -12,7 +12,7 @@ app.config["BLOGGING_URL_PREFIX"] = "/blog"
 app.config["BLOGGING_DISQUS_SITENAME"] = "test"
 app.config["BLOGGING_SITEURL"] = "http://localhost:8000"
 app.config["BLOGGING_SITENAME"] = "My Site"
-app.config["BLOGGING_PERMISSIONS"] = False  # Enable blogger permissions
+app.config["BLOGGING_PERMISSIONS"] = True  # Enable blogger permissions
 
 # extensions
 engine = create_engine('sqlite:////tmp/blog.db')
@@ -35,7 +35,7 @@ def on_identity_loaded(sender, identity):
     identity.user = current_user
     if hasattr(current_user, "id"):
         identity.provides.add(UserNeed(current_user.id))
-    identity.provides.add(RoleNeed("blogger"))
+    #identity.provides.add(RoleNeed("blogger"))
 
 
 @login_manager.user_loader