+++ flask_blogging/processor.py
@@ -6,6 +6,7 @@ import markdown
 from markdown.extensions.meta import MetaExtension
 from flask import url_for
 from flask_login import current_user
+from slugify import slugify
 
 
 class MathJaxPattern(markdown.inlinepatterns.Pattern):
@@ -38,7 +39,7 @@ class PostProcessor(object):
 
     @staticmethod
     def create_slug(title):
-        return "-".join([t.lower() for t in title.split()])
+        return slugify(title)
 
     @classmethod
     def construct_url(cls, post):
@@ -67,22 +68,9 @@ class PostProcessor(object):
         post["slug"] = cls.create_slug(post["title"])
         post["editable"] = cls.is_author(post, current_user)
         post["url"] = cls.construct_url(post)
+        post["priority"] = 0.8
         if render:
             cls.render_text(post)
-        cls.custom_process(post)
-        return
-
-    @classmethod
-    def custom_process(cls, post):
-        """
-        Override this method to add additional processes. The result is that
-        the ``post`` dict is modified or enhanced with newer key value pairs.
-
-        :param post: The post data with values for keys such as title, text,
-         tags etc.
-        :type post: dict
-        """
-        pass
 
     @classmethod
     def all_extensions(cls):
@@ -90,5 +78,5 @@ class PostProcessor(object):
 
     @classmethod
     def set_custom_extensions(cls, extensions):
-        assert type(extensions) == list
-        cls._markdown_extensions.append(cls._custom_extensions)
+        if type(extensions) == list:
+            cls._markdown_extensions.extend(extensions)