+++ flask_blogging/sqlastorage.py
@@ -6,18 +6,20 @@ import logging
 import sqlalchemy as sqla
 import datetime
 from .storage import Storage
+from .signals import sqla_initialized
 
 
 class SQLAStorage(Storage):
     """
     The ``SQLAStorage`` implements the interface specified by the ``Storage``
     class. This  class uses SQLAlchemy to implement storage and retrieval of
-    data from any of the databases supported by SQLAlchemy. This
+    data from any of the databases supported by SQLAlchemy.
     """
     _db = None
     _logger = logging.getLogger("flask-blogging")
 
-    def __init__(self, engine=None, table_prefix="", metadata=None, db=None):
+    def __init__(self, engine=None, table_prefix="", metadata=None, db=None,
+                 bind=None):
         """
         The constructor for the ``SQLAStorage`` class.
 
@@ -31,24 +33,52 @@ class SQLAStorage(Storage):
         :param metadata: (Optional) The SQLAlchemy MetaData object
         :type metadata: object
         :param db: (Optional) The Flask-SQLAlchemy SQLAlchemy object
-        :param db: object
-        :param
+        :type db: object
+        :param bind: (Optional) Reference the database to bind for multiple
+        database scenario with binds
+        :type bind: str
         """
+        self._bind = bind
         if db:
-            self._engine = db.engine
+            self._engine = db.get_engine(db.get_app(), bind=self._bind)
             self._metadata = db.metadata
         else:
             if engine is None:
                 raise ValueError("Both db and engine args cannot be None")
             self._engine = engine
             self._metadata = metadata or sqla.MetaData()
+        self._info = {} if self._bind is None else {"bind_key": self._bind}
         self._table_prefix = table_prefix
         self._metadata.reflect(bind=self._engine)
         self._create_all_tables()
+        sqla_initialized.send(self, engine=self._engine,
+                              table_prefix=self._table_prefix,
+                              meta=self.metadata,
+                              bind=self._bind)
 
     @property
     def metadata(self):
         return self._metadata
+
+    @property
+    def post_table(self):
+        return self._post_table
+
+    @property
+    def tag_table(self):
+        return self._tag_table
+
+    @property
+    def tag_posts_table(self):
+        return self._tag_posts_table
+
+    @property
+    def user_posts_table(self):
+        return self._user_posts_table
+
+    @property
+    def engine(self):
+        return self._engine
 
     def save_post(self, title, text, user_id, tags, draft=False,
                   post_date=None, last_modified_date=None, meta_data=None,
@@ -112,6 +142,7 @@ class SQLAStorage(Storage):
                     if post_id is None else post_id
                 self._save_tags(tags, post_id, conn)
                 self._save_user_post(user_id, post_id, conn)
+
             except Exception as e:
                 self._logger.exception(str(e))
                 post_id = None
@@ -405,6 +436,7 @@ class SQLAStorage(Storage):
         with self._engine.begin() as conn:
             post_table_name = self._table_name("post")
             if not conn.dialect.has_table(conn, post_table_name):
+
                 self._post_table = sqla.Table(
                     post_table_name, self._metadata,
                     sqla.Column("id", sqla.Integer, primary_key=True),
@@ -413,7 +445,8 @@ class SQLAStorage(Storage):
                     sqla.Column("post_date", sqla.DateTime),
                     sqla.Column("last_modified_date", sqla.DateTime),
                     # if 1 then make it a draft
-                    sqla.Column("draft", sqla.SmallInteger, default=0)
+                    sqla.Column("draft", sqla.SmallInteger, default=0),
+                    info=self._info
 
                 )
                 self._logger.debug("Created table with table name %s" %
@@ -435,7 +468,8 @@ class SQLAStorage(Storage):
                     tag_table_name, self._metadata,
                     sqla.Column("id", sqla.Integer, primary_key=True),
                     sqla.Column("text", sqla.String(128), unique=True,
-                                index=True)
+                                index=True),
+                    info=self._info
                 )
                 self._logger.debug("Created table with table name %s" %
                                    tag_table_name)
@@ -466,7 +500,8 @@ class SQLAStorage(Storage):
                                                 onupdate="CASCADE",
                                                 ondelete="CASCADE"),
                                 index=True),
-                    sqla.UniqueConstraint('tag_id', 'post_id', name='uix_1')
+                    sqla.UniqueConstraint('tag_id', 'post_id', name='uix_1'),
+                    info=self._info
                 )
                 self._logger.debug("Created table with table name %s" %
                                    tag_posts_table_name)
@@ -494,7 +529,8 @@ class SQLAStorage(Storage):
                                                 onupdate="CASCADE",
                                                 ondelete="CASCADE"),
                                 index=True),
-                    sqla.UniqueConstraint('user_id', 'post_id', name='uix_2')
+                    sqla.UniqueConstraint('user_id', 'post_id', name='uix_2'),
+                    info=self._info
                 )
                 self._logger.debug("Created table with table name %s" %
                                    user_posts_table_name)